src/main/java/com/booking/BookingController.java
package com.booking;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class BookingController {
    private final BookingRepository bookingRepository;

    public BookingController(BookingRepository bookingRepository) {
        this.bookingRepository = bookingRepository;
    }

    @GetMapping
    public String greetings() {
        return "greetings from booking. !";
    }

    @PostMapping
    public Booking book(@RequestBody Booking booking) {
        return bookingRepository.save(booking);
    }
}
